refactor(tui): move the MultiPaneTui tests next to the code they cover
Seventy-five tests leave the module root for tui::multi_pane, plus the
selection and grid tests for tui::selection and tui::geometry.
