+++ src/tui.rs
@@ -3147,7 +3147,6 @@ fn member_label(peer_id: &[u8], members: &[crate::layout::Member]) -> String {
 #[cfg(test)]
 mod tests {
     use std::{
-        collections::BTreeMap,
         fs,
         net::Ipv4Addr,
         thread,
@@ -3155,14 +3154,14 @@ mod tests {
     };
     use tokio::sync::{mpsc, watch};
 
-    use crossterm::event::{KeyCode, KeyEvent, KeyModifiers, MouseButton, MouseEventKind};
+    use crossterm::event::{KeyCode, KeyEvent, KeyModifiers};
     use portable_pty::PtySize;
-    use ratatui::{Terminal, backend::TestBackend, layout::Rect};
+    use ratatui::layout::Rect;
 
-    use crate::layout::{Axis, LayoutError, NewPanePosition, Node, Tab};
+    use crate::agent_detect::cwd_for_pid;
+    use crate::layout::{Axis, NewPanePosition};
     use crate::lease::{IDLE_AFTER, LeaseManager, LeaseState};
     use crate::screen::HostScreen;
-    use crate::{agent_detect::cwd_for_pid, config::UiTheme};
     use crate::{
         protocol::PaneDescriptor,
         session::{HostSession, SharedLayoutHost, layout_snapshot_from_state},
@@ -3170,429 +3169,12 @@ mod tests {
     };
     use iroh::{Endpoint, RelayMode, endpoint::presets};
 
-    use super::input::keys::{ESC_PREFIX_WINDOW, is_chord_command, is_chord_navigation};
-    use super::multi_pane::keys::CHORD_IDLE_TIMEOUT;
-    use super::test_support::{agent_overlay_tui, agent_row, layout, mouse_protocol, split_layout};
     use super::{
-        AGENT_TOGGLE_WINDOW, ChordMode, HostControlEvent, HostPaneChannels, HostPaneRuntime,
-        KeyHandling, LayoutControlEvent, MultiPaneTui, PaneMouseProtocol, PaneTextSelection,
-        PaneViewState, PendingEscape, RemoteInput, RemoteSubscriptionState, ScreenCell, ShareCopy,
-        SharedLayoutRuntime, SharedLocalPane, UiIntent, grid_for_pane, initial_root_pane_grid,
+        HostControlEvent, HostPaneChannels, HostPaneRuntime, LayoutControlEvent, RemoteInput,
+        RemoteSubscriptionState, SharedLayoutRuntime, SharedLocalPane, UiIntent,
         lease_allows_held_input, member_label, pane_wire_id, reconcile_remote_control_attempt,
-        remote_input_decision, render_multi_pane, selection_text, viewed_screen,
+        remote_input_decision,
     };
-
-    #[test]
-    fn attached_agent_rows_mark_only_unfocused_working_to_idle_transitions_unread() {
-        let snapshot = layout(
-            vec![Tab {
-                tab_id: 1,
-                root: Node::Split {
-                    axis: Axis::LeftRight,
-                    first_share_bps: crate::layout::DEFAULT_FIRST_SHARE_BPS,
-                    first: Box::new(Node::Leaf { pane_id: 1 }),
-                    second: Box::new(Node::Leaf { pane_id: 2 }),
-                },
-                title: None,
-            }],
-            &[(1, 2, 8), (2, 2, 8)],
-        );
-        let mut tui = MultiPaneTui::new(snapshot).expect("valid layout");
-        let working = agent_row(2, 1, 2);
-        let mut idle = working.clone();
-        idle.state = crate::protocol::AgentRosterState::Idle;
-
-        assert_eq!(
-            tui.update_attached_agent_rows(vec![working.clone()]),
-            Vec::<u64>::new()
-        );
-        assert_eq!(tui.update_attached_agent_rows(vec![idle.clone()]), vec![2]);
-        assert!(tui.unread_agent_panes.contains(&2));
-        assert_eq!(
-            tui.update_attached_agent_rows(vec![idle]),
-            Vec::<u64>::new()
-        );
-
-        let mut focused_working = working;
-        focused_working.pane_id = 1;
-        focused_working.pane_ordinal = 1;
-        let mut focused_idle = focused_working.clone();
-        focused_idle.state = crate::protocol::AgentRosterState::Idle;
-        assert_eq!(
-            tui.update_attached_agent_rows(vec![focused_working]),
-            Vec::<u64>::new()
-        );
-        assert_eq!(
-            tui.update_attached_agent_rows(vec![focused_idle]),
-            Vec::<u64>::new()
-        );
-        assert!(!tui.unread_agent_panes.contains(&1));
-    }
-
-    #[test]
-    fn focusing_a_pane_does_not_re_announce_the_same_work_episode() {
-        let snapshot = layout(
-            vec![Tab {
-                tab_id: 1,
-                root: Node::Split {
-                    axis: Axis::LeftRight,
-                    first_share_bps: crate::layout::DEFAULT_FIRST_SHARE_BPS,
-                    first: Box::new(Node::Leaf { pane_id: 1 }),
-                    second: Box::new(Node::Leaf { pane_id: 2 }),
-                },
-                title: None,
-            }],
-            &[(1, 2, 8), (2, 2, 8)],
-        );
-        let mut tui = MultiPaneTui::new(snapshot).expect("valid layout");
-        let working = agent_row(2, 1, 2);
-        let mut idle = working.clone();
-        idle.state = crate::protocol::AgentRosterState::Idle;
-        // A real host sends the `0` sentinel on a non-working row, so the episode a completion
-        // refers to is only ever visible on the working row that preceded it.
-        idle.working_since_unix_ms = 0;
-
-        tui.update_attached_agent_rows(vec![working.clone()]);
-        assert_eq!(tui.update_attached_agent_rows(vec![idle.clone()]), vec![2]);
-
-        // Looking at the pane clears the unread marker. That must not re-arm the sound: the
-        // user has already been told about this episode.
-        tui.set_focus(1, 2).expect("known pane");
-        tui.set_focus(1, 1).expect("known pane");
-        assert!(!tui.unread_agent_panes.contains(&2));
-
-        tui.update_attached_agent_rows(vec![working.clone()]);
-        assert_eq!(
-            tui.update_attached_agent_rows(vec![idle.clone()]),
-            Vec::<u64>::new(),
-            "the same work episode must only be announced once"
-        );
-
-        // A genuinely new working interval is a new episode and does announce again.
-        let mut next_working = working;
-        next_working.working_since_unix_ms += 1;
-        tui.update_attached_agent_rows(vec![next_working]);
-        assert_eq!(tui.update_attached_agent_rows(vec![idle]), vec![2]);
-    }
-
-    #[test]
-    fn attached_agent_rows_preserve_state_and_unread_when_rows_vanish() {
-        let snapshot = layout(
-            vec![Tab {
-                tab_id: 1,
-                root: Node::Split {
-                    axis: Axis::LeftRight,
-                    first_share_bps: crate::layout::DEFAULT_FIRST_SHARE_BPS,
-                    first: Box::new(Node::Leaf { pane_id: 1 }),
-                    second: Box::new(Node::Leaf { pane_id: 2 }),
-                },
-                title: None,
-            }],
-            &[(1, 2, 8), (2, 2, 8)],
-        );
-        let mut tui = MultiPaneTui::new(snapshot).expect("valid layout");
-        let working = agent_row(2, 1, 2);
-        let mut idle = working.clone();
-        idle.state = crate::protocol::AgentRosterState::Idle;
-
-        tui.update_attached_agent_rows(vec![working]);
-        assert_eq!(tui.update_attached_agent_rows(vec![idle]), vec![2]);
-        tui.update_attached_agent_rows(Vec::new());
-        assert_eq!(
-            tui.prior_agent_states[&2],
-            crate::protocol::AgentRosterState::Idle
-        );
-        assert!(tui.unread_agent_panes.contains(&2));
-    }
-
-    #[test]
-    fn focus_routes_clear_unread_agent_markers() {
-        let snapshot = layout(
-            vec![
-                Tab {
-                    tab_id: 1,
-                    root: Node::Leaf { pane_id: 1 },
-                    title: None,
-                },
-                Tab {
-                    tab_id: 2,
-                    root: Node::Leaf { pane_id: 2 },
-                    title: None,
-                },
-            ],
-            &[(1, 2, 8), (2, 2, 8)],
-        );
-        let mut tui = MultiPaneTui::new(snapshot).expect("valid layout");
-
-        tui.unread_agent_panes.insert(2);
-        tui.select_tab(2).expect("known tab");
-        assert!(!tui.unread_agent_panes.contains(&2));
-
-        tui.unread_agent_panes.insert(1);
-        tui.set_focus(1, 1).expect("known pane");
-        assert!(!tui.unread_agent_panes.contains(&1));
-
-        tui.unread_agent_panes.insert(2);
-        assert_eq!(
-            tui.jump_to_agent_pane(2, "test"),
-            vec![UiIntent::FocusPane { pane_id: 2 }]
-        );
-        assert!(!tui.unread_agent_panes.contains(&2));
-    }
-
-    #[test]
-    fn agents_overlay_opens_immediately_and_double_tap_forwards_ctrl_a() {
-        let mut tui = MultiPaneTui::new(layout(
-            vec![Tab {
-                tab_id: 1,
-                root: Node::Leaf { pane_id: 1 },
-                title: None,
-            }],
-            &[(1, 2, 8)],
-        ))
-        .unwrap();
-        let ctrl_a = KeyEvent::new(KeyCode::Char('a'), KeyModifiers::CONTROL);
-        assert_eq!(
-            tui.handle_key(ctrl_a, Rect::new(0, 0, 80, 24)),
-            KeyHandling::Consumed(vec![])
-        );
-        assert!(tui.overlay_open());
-        assert_eq!(
-            tui.handle_key(ctrl_a, Rect::new(0, 0, 80, 24)),
-            KeyHandling::Forward
-        );
-        assert!(!tui.overlay_open());
-        assert_eq!(
-            tui.handle_key(ctrl_a, Rect::new(0, 0, 80, 24)),
-            KeyHandling::Consumed(vec![])
-        );
-        assert!(tui.overlay_open());
-        tui.pending_agent_toggle = Some(Instant::now() - AGENT_TOGGLE_WINDOW);
-        assert!(!tui.expire_agent_toggle(Instant::now()));
-        assert!(tui.overlay_open());
-        assert_eq!(tui.pending_agent_toggle, None);
-        assert_eq!(
-            tui.handle_key(ctrl_a, Rect::new(0, 0, 80, 24)),
-            KeyHandling::Consumed(vec![])
-        );
-        assert!(!tui.overlay_open());
-    }
-
-    #[test]
-    fn agents_overlay_is_modal_and_enter_jumps_to_another_tab() {
-        let snapshot = layout(
-            vec![
-                Tab {
-                    tab_id: 1,
-                    root: Node::Leaf { pane_id: 1 },
-                    title: None,
-                },
-                Tab {
-                    tab_id: 2,
-                    root: Node::Leaf { pane_id: 2 },
-                    title: None,
-                },
-            ],
-            &[(1, 2, 8), (2, 2, 8)],
-        );
-        let mut tui = MultiPaneTui::new(snapshot).unwrap();
-        tui.set_agent_rows(vec![agent_row(2, 2, 1)]);
-        let ctrl_a = KeyEvent::new(KeyCode::Char('a'), KeyModifiers::CONTROL);
-        tui.handle_key(ctrl_a, Rect::new(0, 0, 80, 24));
-        assert_eq!(
-            tui.handle_key(
-                KeyEvent::new(KeyCode::Char('x'), KeyModifiers::NONE),
-                Rect::new(0, 0, 80, 24)
-            ),
-            KeyHandling::Consumed(vec![])
-        );
-        assert_eq!(
-            tui.handle_key(
-                KeyEvent::new(KeyCode::Enter, KeyModifiers::NONE),
-                Rect::new(0, 0, 80, 24)
-            ),
-            KeyHandling::Consumed(vec![UiIntent::FocusPane { pane_id: 2 }])
-        );
-        assert_eq!(tui.current_tab(), 2);
-        assert_eq!(tui.focused_pane(), 2);
-        assert_eq!(tui.pending_agent_toggle, None);
-        assert_eq!(
-            tui.handle_key(ctrl_a, Rect::new(0, 0, 80, 24)),
-            KeyHandling::Consumed(vec![])
-        );
-        assert!(tui.overlay_open());
-        assert_eq!(
-            tui.handle_key(
-                KeyEvent::new(KeyCode::Esc, KeyModifiers::NONE),
-                Rect::new(0, 0, 80, 24)
-            ),
-            KeyHandling::Consumed(vec![])
-        );
-        assert_eq!(tui.pending_agent_toggle, None);
-        assert_eq!(
-            tui.handle_key(ctrl_a, Rect::new(0, 0, 80, 24)),
-            KeyHandling::Consumed(vec![])
-        );
-    }
-
-    #[test]
-    fn agents_overlay_rows_use_chrome_locations_and_sort_by_them() {
-        let snapshot = layout(
-            vec![
-                Tab {
-                    tab_id: 10,
-                    root: Node::Split {
-                        axis: Axis::LeftRight,
-                        first_share_bps: 5_000,
-                        first: Box::new(Node::Leaf { pane_id: 8 }),
-                        second: Box::new(Node::Leaf { pane_id: 6 }),
-                    },
-
-                    title: None,
-                },
-                Tab {
-                    tab_id: 20,
-                    root: Node::Leaf { pane_id: 3 },
-
-                    title: None,
-                },
-            ],
-            &[(8, 2, 8), (6, 2, 8), (3, 2, 8)],
-        );
-        let mut tui = MultiPaneTui::new(snapshot).unwrap();
-
-        tui.set_agent_rows(vec![
-            agent_row(3, 99, 99),
-            agent_row(6, 99, 99),
-            agent_row(8, 99, 99),
-        ]);
-
-        assert_eq!(
-            tui.agent_rows
-                .iter()
-                .map(|row| (row.pane_id, row.tab_ordinal, row.pane_ordinal))
-                .collect::<Vec<_>>(),
-            vec![(8, 1, 1), (6, 1, 2), (3, 2, 1)]
-        );
-        assert!(
-            super::render::agents::format_agent_overlay_card(
-                &tui.agent_rows[2],
-                false,
-                120,
-                0,
-                0,
-                &UiTheme::default(),
-            )[1]
-            .to_string()
-            .contains("Tab #2 · Pane #1")
-        );
-    }
-
-    #[test]
-    fn agents_overlay_click_jumps_to_the_clicked_row_and_ignores_outside_clicks() {
-        let snapshot = layout(
-            vec![
-                Tab {
-                    tab_id: 1,
-                    root: Node::Leaf { pane_id: 1 },
-
-                    title: None,
-                },
-                Tab {
-                    tab_id: 2,
-                    root: Node::Leaf { pane_id: 2 },
-
-                    title: None,
-                },
-            ],
-            &[(1, 2, 8), (2, 2, 8)],
-        );
-        let mut tui = MultiPaneTui::new(snapshot).unwrap();
-        tui.set_agent_rows(vec![agent_row(1, 1, 1), agent_row(2, 2, 1)]);
-        let area = Rect::new(0, 0, 80, 24);
-        let ctrl_a = KeyEvent::new(KeyCode::Char('a'), KeyModifiers::CONTROL);
-        tui.handle_key(ctrl_a, area);
-
-        assert_eq!(tui.handle_agent_overlay_click(0, 0, area), Vec::new());
-        assert!(tui.overlay_open());
-        let inner = super::render::agents::agents_overlay_inner(area);
-        let content = super::render::agents::agents_overlay_content(area);
-        let help = super::render::agents::agents_overlay_help(area).expect("help row");
-        assert_eq!(
-            tui.agent_overlay_row_at(content.x.saturating_sub(1), content.y, area),
-            None
-        );
-        assert_eq!(tui.agent_overlay_row_at(content.x, inner.y, area), None);
-        assert_eq!(tui.agent_overlay_row_at(help.x, help.y, area), None);
-        assert_eq!(
-            tui.agent_overlay_row_at(content.x, content.y, area),
-            Some(1)
-        );
-        assert_eq!(
-            tui.agent_overlay_row_at(content.x, content.y.saturating_add(1), area),
-            Some(1)
-        );
-        assert_eq!(
-            tui.agent_overlay_row_at(content.x, content.y.saturating_add(3), area),
-            Some(2)
-        );
-        assert_eq!(
-            tui.agent_overlay_row_at(content.x, content.y.saturating_add(4), area),
-            Some(2)
-        );
-        assert_eq!(
-            tui.handle_agent_overlay_click(content.x, content.y.saturating_add(3), area),
-            vec![UiIntent::FocusPane { pane_id: 2 }]
-        );
-        assert!(!tui.overlay_open());
-        assert_eq!(tui.current_tab(), 2);
-        assert_eq!(tui.focused_pane(), 2);
-        assert_eq!(tui.pending_agent_toggle, None);
-        assert_eq!(tui.handle_key(ctrl_a, area), KeyHandling::Consumed(vec![]));
-    }
-
-    #[test]
-    fn agents_overlay_hit_test_accounts_for_terminal_line_scroll() {
-        let area = Rect::new(0, 0, 80, 8);
-        let mut tui = agent_overlay_tui(3);
-        let content = super::render::agents::agents_overlay_content(area);
-
-        assert!(tui.scroll_agent_overlay(area, false));
-        assert_eq!(tui.agent_overlay_scroll_line, 3);
-        assert_eq!(
-            tui.agent_overlay_row_at(content.x, content.y, area),
-            Some(2)
-        );
-        assert_eq!(
-            tui.agent_overlay_row_at(content.x, content.y.saturating_add(2), area),
-            None
-        );
-    }
-
-    #[test]
-    fn agents_overlay_selection_auto_scrolls_to_keep_cards_visible() {
-        let area = Rect::new(0, 0, 80, 8);
-        let mut tui = agent_overlay_tui(4);
-
-        tui.handle_key(KeyEvent::new(KeyCode::Down, KeyModifiers::NONE), area);
-        assert_eq!(tui.agent_selected_pane, Some(2));
-        assert_eq!(tui.agent_overlay_scroll_line, 3);
-        tui.handle_key(KeyEvent::new(KeyCode::Down, KeyModifiers::NONE), area);
-        assert_eq!(tui.agent_selected_pane, Some(3));
-        assert_eq!(tui.agent_overlay_scroll_line, 6);
-    }
-
-    #[test]
-    fn agents_overlay_scroll_clamps_when_roster_shrinks() {
-        let area = Rect::new(0, 0, 80, 8);
-        let mut tui = agent_overlay_tui(4);
-
-        assert!(tui.scroll_agent_overlay(area, false));
-        assert!(tui.scroll_agent_overlay(area, false));
-        assert_eq!(tui.agent_overlay_scroll_line, 6);
-        tui.set_agent_rows(vec![agent_row(1, 1, 1)]);
-        assert_eq!(tui.agent_overlay_scroll_line, 0);
-    }
 
     #[test]
     fn failed_remote_subscription_retries_with_bounded_backoff_and_commit_nudge() {
@@ -3980,35 +3562,6 @@ mod tests {
     }
 
     #[test]
-    fn pane_view_activity_transition_reports_a_redraw() {
-        let snapshot = layout(
-            vec![Tab {
-                tab_id: 1,
-                root: Node::Leaf { pane_id: 1 },
-
-                title: None,
-            }],
-            &[(1, 1, 1)],
-        );
-        let mut tui = MultiPaneTui::new(snapshot).expect("layout");
-        let active = PaneViewState {
-            ready: true,
-            controller_peer_id: Some(b"controller".to_vec()),
-            controller_active: true,
-            scrollback: 0,
-        };
-        assert!(tui.set_pane_view(1, active.clone()));
-        assert!(!tui.set_pane_view(1, active));
-        assert!(tui.set_pane_view(
-            1,
-            PaneViewState {
-                controller_active: false,
-                ..tui.pane_view(1).expect("pane view").clone()
-            },
-        ));
-    }
-
-    #[test]
     fn rejected_busy_claim_republishes_lease_and_idle_claim_succeeds() {
         let owner = b"owner".to_vec();
         let requester = b"guest".to_vec();
@@ -4339,1176 +3892,6 @@ mod tests {
     }
 
     #[test]
-    fn content_drag_keeps_the_selection_path() {
-        let mut tui = MultiPaneTui::new(split_layout()).expect("layout");
-        let area = Rect::new(0, 0, 80, 24);
-        assert!(!tui.begin_resize_drag(2, 3, area));
-        assert!(tui.resize_drag.is_none());
-        assert!(tui.begin_selection_at(2, 3, area));
-        assert!(tui.extend_selection_at(3, 3, area));
-        assert!(tui.end_selection_drag());
-        assert!(tui.selection().is_some());
-    }
-
-    fn left_mouse(kind: MouseEventKind, column: u16, row: u16) -> crossterm::event::MouseEvent {
-        crossterm::event::MouseEvent {
-            kind,
-            column,
-            row,
-            modifiers: KeyModifiers::NONE,
-        }
-    }
-
-    fn reporting_child() -> PaneMouseProtocol {
-        mouse_protocol(
-            vt100::MouseProtocolMode::ButtonMotion,
-            vt100::MouseProtocolEncoding::Sgr,
-        )
-    }
-
-    #[test]
-    fn clicking_a_focused_reporting_pane_forwards_instead_of_selecting() {
-        let mut tui = MultiPaneTui::new(split_layout()).expect("layout");
-        let area = Rect::new(0, 0, 80, 24);
-
-        let handling = tui.handle_mouse(
-            left_mouse(MouseEventKind::Down(MouseButton::Left), 2, 3),
-            area,
-            reporting_child(),
-        );
-
-        assert_eq!(handling.forward_bytes, Some(b"\x1b[<0;2;2M".to_vec()));
-        assert!(handling.intents.is_empty());
-        assert_eq!(tui.selection(), None);
-    }
-
-    #[test]
-    fn a_forwarded_press_keeps_the_drag_and_release_away_from_selection() {
-        let mut tui = MultiPaneTui::new(split_layout()).expect("layout");
-        let area = Rect::new(0, 0, 80, 24);
-        let protocol = reporting_child();
-
-        tui.handle_mouse(
-            left_mouse(MouseEventKind::Down(MouseButton::Left), 2, 3),
-            area,
-            protocol,
-        );
-        let drag = tui.handle_mouse(
-            left_mouse(MouseEventKind::Drag(MouseButton::Left), 4, 3),
-            area,
-            protocol,
-        );
-        let up = tui.handle_mouse(
-            left_mouse(MouseEventKind::Up(MouseButton::Left), 4, 3),
-            area,
-            protocol,
-        );
-
-        assert_eq!(drag.forward_bytes, Some(b"\x1b[<32;4;2M".to_vec()));
-        assert_eq!(up.forward_bytes, Some(b"\x1b[<0;4;2m".to_vec()));
-        assert!(!up.copy_selection_requested);
-        assert_eq!(tui.selection(), None);
-    }
-
-    #[test]
-    fn a_drag_that_leaves_the_pane_still_reports_inside_its_edge() {
-        let mut tui = MultiPaneTui::new(split_layout()).expect("layout");
-        let area = Rect::new(0, 0, 80, 24);
-        let protocol = reporting_child();
-
-        tui.handle_mouse(
-            left_mouse(MouseEventKind::Down(MouseButton::Left), 2, 3),
-            area,
-            protocol,
-        );
-        let drag = tui.handle_mouse(
-            left_mouse(MouseEventKind::Drag(MouseButton::Left), 70, 20),
-            area,
-            protocol,
-        );
-
-        // The 10x4 grid clamps the report to its last cell rather than dropping it.
-        assert_eq!(drag.forward_bytes, Some(b"\x1b[<32;10;4M".to_vec()));
-    }
-
-    #[test]
-    fn shift_click_selects_even_when_the_child_reports_mouse() {
-        let mut tui = MultiPaneTui::new(split_layout()).expect("layout");
-        let area = Rect::new(0, 0, 80, 24);
-        let shift_down = crossterm::event::MouseEvent {
-            kind: MouseEventKind::Down(MouseButton::Left),
-            column: 2,
-            row: 3,
-            modifiers: KeyModifiers::SHIFT,
-        };
-
-        let handling = tui.handle_mouse(shift_down, area, reporting_child());
-
-        assert_eq!(handling.forward_bytes, None);
-        assert!(tui.selection_dragging);
-    }
-
-    #[test]
-    fn clicking_an_unfocused_reporting_pane_only_moves_focus() {
-        let mut tui = MultiPaneTui::new(split_layout()).expect("layout");
-        let area = Rect::new(0, 0, 80, 24);
-        assert_eq!(tui.focused_pane(), 1);
-
-        let handling = tui.handle_mouse(
-            left_mouse(MouseEventKind::Down(MouseButton::Left), 45, 3),
-            area,
-            reporting_child(),
-        );
-
-        assert_eq!(handling.forward_bytes, None);
-        assert_eq!(tui.focused_pane(), 2);
-        assert!(matches!(
-            handling.intents.as_slice(),
-            [UiIntent::FocusPane { pane_id: 2 }]
-        ));
-    }
-
-    #[test]
-    fn a_scrolled_back_pane_selects_history_instead_of_forwarding() {
-        let mut tui = MultiPaneTui::new(split_layout()).expect("layout");
-        let area = Rect::new(0, 0, 80, 24);
-        assert!(tui.set_pane_scrollback_offset(1, 5));
-
-        let handling = tui.handle_mouse(
-            left_mouse(MouseEventKind::Down(MouseButton::Left), 2, 3),
-            area,
-            reporting_child(),
-        );
-
-        assert_eq!(handling.forward_bytes, None);
-        assert!(tui.selection_dragging);
-    }
-
-    #[test]
-    fn the_wheel_reaches_a_reporting_child_over_its_own_pane() {
-        let mut tui = MultiPaneTui::new(split_layout()).expect("layout");
-        let area = Rect::new(0, 0, 80, 24);
-
-        let handling = tui.handle_mouse(
-            left_mouse(MouseEventKind::ScrollUp, 2, 3),
-            area,
-            reporting_child(),
-        );
-
-        assert_eq!(handling.forward_bytes, Some(b"\x1b[<64;2;2M".to_vec()));
-    }
-
-    #[test]
-    fn the_wheel_stays_local_over_a_pane_the_reporting_child_does_not_own() {
-        let mut tui = MultiPaneTui::new(split_layout()).expect("layout");
-        let area = Rect::new(0, 0, 80, 24);
-
-        let handling = tui.handle_mouse(
-            left_mouse(MouseEventKind::ScrollDown, 45, 3),
-            area,
-            reporting_child(),
-        );
-
-        assert_eq!(handling.forward_bytes, None);
-    }
-
-    #[test]
-    fn the_wheel_stays_local_while_a_reporting_pane_shows_history() {
-        let mut tui = MultiPaneTui::new(split_layout()).expect("layout");
-        let area = Rect::new(0, 0, 80, 24);
-        assert!(tui.set_pane_scrollback_offset(1, 3));
-
-        let handling = tui.handle_mouse(
-            left_mouse(MouseEventKind::ScrollDown, 2, 3),
-            area,
-            reporting_child(),
-        );
-
-        assert_eq!(handling.forward_bytes, None);
-    }
-
-    #[test]
-    fn a_reporting_child_never_takes_a_border_drag_from_the_mux() {
-        let mut tui = MultiPaneTui::new(split_layout()).expect("layout");
-        let area = Rect::new(0, 0, 80, 24);
-        let protocol = reporting_child();
-
-        let down = tui.handle_mouse(
-            left_mouse(MouseEventKind::Down(MouseButton::Left), 39, 5),
-            area,
-            protocol,
-        );
-        tui.handle_mouse(
-            left_mouse(MouseEventKind::Drag(MouseButton::Left), 49, 5),
-            area,
-            protocol,
-        );
-        let up = tui.handle_mouse(
-            left_mouse(MouseEventKind::Up(MouseButton::Left), 49, 5),
-            area,
-            protocol,
-        );
-
-        assert_eq!(down.forward_bytes, None);
-        assert!(matches!(
-            up.intents.as_slice(),
-            [UiIntent::SetSplitRatio { .. }]
-        ));
-    }
-
-    #[test]
-    fn mouse_up_requests_copy_only_for_a_nonempty_selection_without_resize_commit() {
-        let mut tui = MultiPaneTui::new(split_layout()).expect("layout");
-        let area = Rect::new(0, 0, 80, 24);
-        let down = crossterm::event::MouseEvent {
-            kind: MouseEventKind::Down(MouseButton::Left),
-            column: 2,
-            row: 3,
-            modifiers: KeyModifiers::NONE,
-        };
-        let drag = crossterm::event::MouseEvent {
-            kind: MouseEventKind::Drag(MouseButton::Left),
-            column: 3,
-            row: 3,
-            modifiers: KeyModifiers::NONE,
-        };
-        let up = crossterm::event::MouseEvent {
-            kind: MouseEventKind::Up(MouseButton::Left),
-            column: 3,
-            row: 3,
-            modifiers: KeyModifiers::NONE,
-        };
-
-        assert!(
-            tui.handle_mouse(down, area, PaneMouseProtocol::default())
-                .intents
-                .is_empty()
-        );
-        tui.handle_mouse(drag, area, PaneMouseProtocol::default());
-        let handling = tui.handle_mouse(up, area, PaneMouseProtocol::default());
-        assert!(handling.intents.is_empty());
-        assert!(handling.copy_selection_requested);
-
-        let resize_down = crossterm::event::MouseEvent {
-            kind: MouseEventKind::Down(MouseButton::Left),
-            column: 39,
-            row: 5,
-            modifiers: KeyModifiers::NONE,
-        };
-        let resize_drag = crossterm::event::MouseEvent {
-            kind: MouseEventKind::Drag(MouseButton::Left),
-            column: 49,
-            row: 5,
-            modifiers: KeyModifiers::NONE,
-        };
-        let resize_up = crossterm::event::MouseEvent {
-            kind: MouseEventKind::Up(MouseButton::Left),
-            column: 49,
-            row: 5,
-            modifiers: KeyModifiers::NONE,
-        };
-        tui.handle_mouse(resize_down, area, PaneMouseProtocol::default());
-        tui.handle_mouse(resize_drag, area, PaneMouseProtocol::default());
-        let handling = tui.handle_mouse(resize_up, area, PaneMouseProtocol::default());
-        assert!(matches!(
-            handling.intents.as_slice(),
-            [UiIntent::SetSplitRatio { .. }]
-        ));
-        assert!(!handling.copy_selection_requested);
-    }
-
-    #[test]
-    fn dragging_a_shared_vertical_border_resizes_its_split() {
-        let mut tui = MultiPaneTui::new(split_layout()).expect("layout");
-        let area = Rect::new(0, 0, 80, 24);
-        assert!(tui.begin_resize_drag(39, 5, area));
-        assert!(tui.extend_resize_drag(49, 5));
-        assert!(matches!(
-            tui.end_resize_drag(49, 5),
-            Some(UiIntent::SetSplitRatio {
-                pane_id: 1,
-                axis: Axis::LeftRight,
-                first_share_bps: 6_250,
-                base_revision: 1,
-            })
-        ));
-        assert!(tui.end_resize_drag(49, 5).is_none());
-
-        assert!(tui.begin_resize_drag(40, 5, area));
-        assert!(tui.extend_resize_drag(50, 5));
-        assert!(matches!(
-            tui.end_resize_drag(50, 5),
-            Some(UiIntent::SetSplitRatio {
-                pane_id: 2,
-                axis: Axis::LeftRight,
-                first_share_bps: 6_250,
-                base_revision: 1,
-            })
-        ));
-    }
-
-    #[test]
-    fn resize_drag_previews_geometry_without_mutating_the_snapshot() {
-        let mut tui = MultiPaneTui::new(split_layout()).expect("layout");
-        let area = Rect::new(0, 0, 80, 24);
-        let snapshot = tui.snapshot().clone();
-        let initial = tui.geometry(area);
-
-        assert!(tui.begin_resize_drag(39, 5, area));
-        assert!(tui.extend_resize_drag(49, 5));
-
-        let preview = tui.geometry(area);
-        assert_eq!(initial.panes[&1], Rect::new(0, 1, 40, 22));
-        assert_eq!(preview.panes[&1], Rect::new(0, 1, 50, 22));
-        assert_eq!(tui.snapshot(), &snapshot);
-        assert!(matches!(
-            tui.end_resize_drag(49, 5),
-            Some(UiIntent::SetSplitRatio {
-                pane_id: 1,
-                axis: Axis::LeftRight,
-                first_share_bps: 6_250,
-                base_revision: 1,
-            })
-        ));
-        assert_eq!(tui.snapshot(), &snapshot);
-        assert_eq!(tui.geometry(area), initial);
-    }
-
-    #[test]
-    fn same_revision_snapshot_preserves_resize_preview_and_selection() {
-        let mut tui = MultiPaneTui::new(split_layout()).expect("layout");
-        let area = Rect::new(0, 0, 80, 24);
-        let initial = tui.geometry(area);
-
-        assert!(tui.begin_resize_drag(39, 5, area));
-        assert!(tui.extend_resize_drag(49, 5));
-        assert!(tui.begin_selection_at(2, 3, area));
-        assert!(tui.extend_selection_at(3, 3, area));
-        assert_ne!(tui.geometry(area), initial);
-
-        tui.apply_snapshot(tui.snapshot().clone())
-            .expect("valid snapshot");
-        assert!(tui.resize_drag.is_some());
-        assert!(tui.selection().is_some());
-        assert_ne!(tui.geometry(area), initial);
-    }
-
-    #[test]
-    fn same_revision_different_snapshot_is_rejected() {
-        let mut tui = MultiPaneTui::new(split_layout()).expect("layout");
-        let mut conflicting = tui.snapshot().clone();
-        conflicting.tabs[0].title = Some("conflict".into());
-
-        assert_eq!(
-            tui.apply_snapshot(conflicting),
-            Err(LayoutError::ConflictingSnapshotRevision { revision: 1 })
-        );
-    }
-
-    #[test]
-    fn higher_revision_snapshot_cancels_resize_preview_and_selection() {
-        let mut tui = MultiPaneTui::new(split_layout()).expect("layout");
-        let area = Rect::new(0, 0, 80, 24);
-        assert!(tui.begin_resize_drag(39, 5, area));
-        assert!(tui.extend_resize_drag(49, 5));
-        assert!(tui.begin_selection_at(2, 3, area));
-        assert!(tui.extend_selection_at(3, 3, area));
-        let mut newer = tui.snapshot().clone();
-        newer.revision += 1;
-
-        tui.apply_snapshot(newer).expect("valid snapshot");
-        assert!(tui.resize_drag.is_none());
-        assert!(tui.selection().is_none());
-    }
-
-    #[test]
-    fn dragging_a_shared_horizontal_border_resizes_its_split() {
-        let mut tui = MultiPaneTui::new(split_layout()).expect("layout");
-        let area = Rect::new(0, 0, 80, 24);
-        let initial = tui.geometry(area);
-        assert!(tui.begin_resize_drag(60, 11, area));
-        assert!(tui.extend_resize_drag(60, 11));
-        assert_eq!(tui.geometry(area), initial);
-        assert!(tui.extend_resize_drag(60, 16));
-        assert_eq!(tui.geometry(area).panes[&2], Rect::new(40, 1, 40, 15));
-        assert!(matches!(
-            tui.end_resize_drag(60, 16),
-            Some(UiIntent::SetSplitRatio {
-                pane_id: 2,
-                axis: Axis::TopBottom,
-                first_share_bps: 7_272,
-                base_revision: 1,
-            })
-        ));
-    }
-
-    #[test]
-    fn rename_prompt_captures_chord_target_edits_and_consumes_modal_input() {
-        let area = Rect::new(0, 0, 80, 24);
-        let mut tui = MultiPaneTui::new(layout(
-            vec![Tab {
-                tab_id: 1,
-                root: Node::Leaf { pane_id: 1 },
-                title: None,
-            }],
-            &[(1, 2, 8)],
-        ))
-        .unwrap();
-        tui.snapshot.panes.get_mut(&1).unwrap().title = Some(String::from("old"));
-
-        assert_eq!(
-            tui.handle_key(
-                KeyEvent::new(KeyCode::Char('p'), KeyModifiers::CONTROL),
-                area
-            ),
-            KeyHandling::Consumed(vec![])
-        );
-        assert_eq!(
-            tui.handle_key(KeyEvent::new(KeyCode::Char('e'), KeyModifiers::NONE), area),
-            KeyHandling::Consumed(vec![])
-        );
-        assert_eq!(tui.chord_mode(), ChordMode::None);
-        assert!(matches!(
-            &tui.modal,
-            super::ModalState::Rename(prompt) if prompt.value == "old"
-        ));
-        assert_eq!(
-            tui.handle_key(
-                KeyEvent::new(KeyCode::Char('🙂'), KeyModifiers::SHIFT),
-                area
-            ),
-            KeyHandling::Consumed(vec![])
-        );
-        assert_eq!(
-            tui.handle_key(KeyEvent::new(KeyCode::Backspace, KeyModifiers::NONE), area),
-            KeyHandling::Consumed(vec![])
-        );
-        assert_eq!(
-            tui.handle_key(KeyEvent::new(KeyCode::Enter, KeyModifiers::NONE), area),
-            KeyHandling::Consumed(vec![UiIntent::RenamePane {
-                pane_id: 1,
-                title: String::from("old"),
-            }])
-        );
-        assert!(!tui.overlay_open());
-    }
-
-    #[test]
-    fn rename_prompt_rejects_invalid_input_and_ctrl_q_wins() {
-        let area = Rect::new(0, 0, 80, 24);
-        let mut tui = MultiPaneTui::new(layout(
-            vec![Tab {
-                tab_id: 1,
-                root: Node::Leaf { pane_id: 1 },
-                title: None,
-            }],
-            &[(1, 2, 8)],
-        ))
-        .unwrap();
-        tui.open_rename(super::RenameTarget::Tab(1));
-        for character in "x".repeat(33).chars() {
-            tui.handle_key(
-                KeyEvent::new(KeyCode::Char(character), KeyModifiers::NONE),
-                area,
-            );
-        }
-        assert_eq!(
-            tui.handle_key(KeyEvent::new(KeyCode::Enter, KeyModifiers::NONE), area),
-            KeyHandling::Consumed(vec![])
-        );
-        assert!(matches!(tui.modal, super::ModalState::Rename(_)));
-        assert_eq!(
-            tui.handle_key(
-                KeyEvent::new(KeyCode::Char('q'), KeyModifiers::CONTROL),
-                area
-            ),
-            KeyHandling::Quit
-        );
-        assert!(matches!(tui.modal, super::ModalState::None));
-    }
-
-    #[test]
-    fn multi_pane_tab_delete_requires_confirmation_and_blocks_pane_input() {
-        let area = Rect::new(0, 0, 80, 24);
-        let mut tui = MultiPaneTui::new(split_layout()).expect("layout");
-
-        let _ = tui.handle_key(
-            KeyEvent::new(KeyCode::Char('t'), KeyModifiers::CONTROL),
-            area,
-        );
-        assert_eq!(
-            tui.handle_key(KeyEvent::new(KeyCode::Char('x'), KeyModifiers::NONE), area),
-            KeyHandling::Consumed(vec![])
-        );
-        assert_eq!(tui.chord_mode(), ChordMode::None);
-        assert!(tui.modal_open());
-        assert!(matches!(
-            &tui.modal,
-            super::ModalState::ConfirmDeleteTab {
-                tab_id: 1,
-                pane_count: 3
-            }
-        ));
-        let mut terminal = Terminal::new(TestBackend::new(80, 24)).expect("terminal");
-        terminal
-            .draw(|frame| render_multi_pane(frame, &tui, &BTreeMap::new()))
-            .expect("render");
-        let mut rendered = String::new();
-        for row in 0..24 {
-            for column in 0..80 {
-                rendered.push_str(terminal.backend().buffer()[(column, row)].symbol());
-            }
-        }
-        assert!(rendered.contains("Delete tab?"));
-        assert!(rendered.contains("3 panes"));
-        assert!(rendered.contains("Enter/y yes · Esc/n no"));
-
-        let mouse = crossterm::event::MouseEvent {
-            kind: MouseEventKind::Down(MouseButton::Left),
-            column: 60,
-            row: 17,
-            modifiers: KeyModifiers::NONE,
-        };
-        assert_eq!(
-            tui.handle_mouse(mouse, area, PaneMouseProtocol::default()),
-            super::MouseHandling::default()
-        );
-        assert_eq!(tui.focused_pane(), 1);
-        assert!(!tui.scroll_mouse_pane(60, 17, area, 10, true));
-        assert!(tui.resize_drag.is_none());
-        assert!(tui.selection.is_none());
-        assert_eq!(
-            tui.handle_key(KeyEvent::new(KeyCode::Char('z'), KeyModifiers::NONE), area),
-            KeyHandling::Consumed(vec![])
-        );
-        assert!(tui.modal_open());
-
-        assert_eq!(
-            tui.handle_key(KeyEvent::new(KeyCode::Char('n'), KeyModifiers::NONE), area),
-            KeyHandling::Consumed(vec![])
-        );
-        assert!(!tui.modal_open());
-
-        let _ = tui.handle_key(
-            KeyEvent::new(KeyCode::Char('t'), KeyModifiers::CONTROL),
-            area,
-        );
-        let _ = tui.handle_key(KeyEvent::new(KeyCode::Char('x'), KeyModifiers::NONE), area);
-        assert_eq!(
-            tui.handle_key(KeyEvent::new(KeyCode::Enter, KeyModifiers::NONE), area),
-            KeyHandling::Consumed(vec![UiIntent::DeleteTab { tab_id: 1 }])
-        );
-        assert!(!tui.modal_open());
-    }
-
-    #[test]
-    fn single_pane_tab_delete_is_immediate() {
-        let area = Rect::new(0, 0, 80, 24);
-        let mut tui = MultiPaneTui::new(layout(
-            vec![Tab {
-                tab_id: 1,
-                root: Node::Leaf { pane_id: 1 },
-                title: None,
-            }],
-            &[(1, 2, 8)],
-        ))
-        .expect("layout");
-
-        let _ = tui.handle_key(
-            KeyEvent::new(KeyCode::Char('t'), KeyModifiers::CONTROL),
-            area,
-        );
-        assert_eq!(
-            tui.handle_key(KeyEvent::new(KeyCode::Char('x'), KeyModifiers::NONE), area),
-            KeyHandling::Consumed(vec![UiIntent::DeleteTab { tab_id: 1 }])
-        );
-        assert!(!tui.modal_open());
-    }
-
-    #[test]
-    fn multi_pane_geometry_recursively_splits_the_content_area() {
-        let tui = MultiPaneTui::new(split_layout()).expect("valid layout");
-        let geometry = tui.geometry(Rect::new(0, 0, 80, 24));
-
-        assert_eq!(geometry.tab_bar, Rect::new(0, 0, 80, 1));
-        assert_eq!(geometry.footer, Rect::new(0, 23, 80, 1));
-        assert_eq!(geometry.content, Rect::new(0, 1, 80, 22));
-        assert_eq!(geometry.panes[&1], Rect::new(0, 1, 40, 22));
-        assert_eq!(geometry.panes[&2], Rect::new(40, 1, 40, 11));
-        assert_eq!(geometry.panes[&3], Rect::new(40, 12, 40, 11));
-    }
-
-    #[test]
-    fn mouse_hit_testing_focuses_the_clicked_pane_and_ignores_misses() {
-        let mut tui = MultiPaneTui::new(split_layout()).expect("valid layout");
-        let area = Rect::new(0, 0, 80, 24);
-        let _ = tui.handle_key(
-            KeyEvent::new(KeyCode::Char('p'), KeyModifiers::CONTROL),
-            area,
-        );
-
-        assert!(tui.focus_pane_at(60, 17, area));
-        assert_eq!(tui.focused_pane(), 3);
-        assert_eq!(tui.chord_mode(), ChordMode::Pane);
-        assert!(!tui.focus_pane_at(10, 0, area));
-        assert_eq!(tui.focused_pane(), 3);
-    }
-
-    #[test]
-    fn hovering_an_unfocused_pane_does_not_move_focus() {
-        let mut tui = MultiPaneTui::new(split_layout()).expect("valid layout");
-        let area = Rect::new(0, 0, 80, 24);
-
-        assert!(tui.hover_pane_at(60, 17, area));
-        assert_eq!(tui.hovered_pane, Some(3));
-        assert_eq!(tui.focused_pane(), 1);
-        assert!(tui.hover_pane_at(10, 0, area));
-        assert_eq!(tui.hovered_pane, None);
-        assert_eq!(tui.focused_pane(), 1);
-    }
-
-    #[test]
-    fn selection_hit_testing_uses_the_bordered_pane_content_area() {
-        let tui = MultiPaneTui::new(layout(
-            vec![Tab {
-                tab_id: 1,
-                root: Node::Leaf { pane_id: 1 },
-
-                title: None,
-            }],
-            &[(1, 19, 78)],
-        ))
-        .expect("valid layout");
-        let area = Rect::new(0, 0, 80, 24);
-
-        assert_eq!(tui.screen_cell_at(0, 1, area), None);
-        assert_eq!(
-            tui.screen_cell_at(1, 2, area),
-            Some((1, ScreenCell { row: 0, col: 0 }))
-        );
-    }
-
-    #[test]
-    fn mouse_wheel_adjusts_the_hovered_pane_scrollback_with_clamping() {
-        let mut tui = MultiPaneTui::new(split_layout()).expect("valid layout");
-        let area = Rect::new(0, 0, 80, 24);
-        let pane_id = tui.pane_at_or_focused(60, 17, area);
-        assert_eq!(pane_id, 3);
-
-        assert!(tui.scroll_pane(pane_id, 10, true));
-        assert_eq!(tui.pane_view(pane_id).expect("pane view").scrollback, 3);
-        assert!(tui.scroll_pane(pane_id, 4, true));
-        assert_eq!(tui.pane_view(pane_id).expect("pane view").scrollback, 4);
-        assert!(!tui.scroll_pane(pane_id, 4, true));
-        assert!(tui.scroll_pane(pane_id, 4, false));
-        assert_eq!(tui.pane_view(pane_id).expect("pane view").scrollback, 1);
-        assert!(tui.scroll_pane(pane_id, 4, false));
-        assert!(!tui.scroll_pane(pane_id, 4, false));
-        assert_eq!(tui.pane_view(pane_id).expect("pane view").scrollback, 0);
-    }
-
-    #[test]
-    fn input_resets_the_pane_scrollback_to_the_live_edge() {
-        let mut tui = MultiPaneTui::new(split_layout()).expect("valid layout");
-        assert!(tui.scroll_pane(1, 4, true));
-        assert!(tui.reset_scrollback(1));
-        assert_eq!(tui.pane_view(1).expect("pane view").scrollback, 0);
-        assert!(!tui.reset_scrollback(1));
-    }
-
-    #[test]
-    fn appended_local_history_pins_a_scrolled_back_viewport() {
-        let mut tui = MultiPaneTui::new(split_layout()).expect("valid layout");
-        assert!(tui.scroll_pane(1, 10, true));
-        assert!(tui.scroll_pane(1, 10, true));
-        assert!(tui.pin_scrollback_after_output(1, 3, 10));
-        assert_eq!(tui.pane_view(1).expect("pane view").scrollback, 9);
-        tui.reset_scrollback(1);
-        assert!(!tui.pin_scrollback_after_output(1, 3, 10));
-    }
-
-    #[test]
-    fn selection_text_uses_stream_semantics_in_document_order() {
-        let mut parser = vt100::Parser::new(3, 4, 0);
-        parser.process(b"abcd\r\nefgh\r\nijkl");
-        let selection = PaneTextSelection {
-            pane_id: 1,
-            anchor: ScreenCell { row: 2, col: 1 },
-            cursor: ScreenCell { row: 0, col: 2 },
-        };
-
-        assert_eq!(
-            selection_text(parser.screen(), selection),
-            Some("cd\nefgh\nij".to_owned())
-        );
-    }
-
-    #[test]
-    fn selection_text_uses_the_scrolled_view_cells() {
-        let mut parser = vt100::Parser::new(1, 3, 10);
-        parser.process(b"one\r\ntwo");
-        let screen = viewed_screen(parser.screen(), 1);
-        let selection = PaneTextSelection {
-            pane_id: 1,
-            anchor: ScreenCell { row: 0, col: 0 },
-            cursor: ScreenCell { row: 0, col: 1 },
-        };
-
-        assert_eq!(selection_text(&screen, selection), Some("on".to_owned()));
-    }
-
-    #[test]
-    fn empty_selection_has_no_clipboard_text() {
-        let parser = vt100::Parser::new(1, 1, 0);
-        let selection = PaneTextSelection {
-            pane_id: 1,
-            anchor: ScreenCell { row: 0, col: 0 },
-            cursor: ScreenCell { row: 0, col: 0 },
-        };
-
-        assert_eq!(selection_text(parser.screen(), selection), None);
-    }
-
-    #[test]
-    fn tab_hit_testing_switches_the_clicked_rendered_label_without_exiting_tab_mode() {
-        let mut tui = MultiPaneTui::new(layout(
-            vec![
-                Tab {
-                    tab_id: 1,
-                    root: Node::Leaf { pane_id: 1 },
-
-                    title: None,
-                },
-                Tab {
-                    tab_id: 2,
-                    root: Node::Leaf { pane_id: 2 },
-
-                    title: None,
-                },
-            ],
-            &[(1, 2, 2), (2, 2, 2)],
-        ))
-        .expect("valid layout");
-        let area = Rect::new(0, 0, 32, 8);
-
-        let _ = tui.handle_key(
-            KeyEvent::new(KeyCode::Char('t'), KeyModifiers::CONTROL),
-            area,
-        );
-
-        let second_tab = tui.geometry(area).tab_labels[&2];
-        assert_eq!(
-            tui.switch_tab_at(second_tab.x, 0, area),
-            Some(UiIntent::SwitchTab { tab_id: 2 })
-        );
-        assert_eq!(tui.current_tab(), 2);
-        assert_eq!(tui.focused_pane(), 2);
-        assert_eq!(tui.chord_mode(), ChordMode::Tab);
-        assert_eq!(tui.switch_tab_at(0, 0, area), None);
-    }
-
-    #[test]
-    fn tiny_terminal_geometry_stays_in_bounds() {
-        let tui = MultiPaneTui::new(split_layout()).expect("valid layout");
-        let geometry = tui.geometry(Rect::new(u16::MAX, u16::MAX, 1, 1));
-
-        assert_eq!(geometry.tab_bar, Rect::new(u16::MAX, u16::MAX, 1, 1));
-        assert_eq!(geometry.footer, Rect::new(u16::MAX, u16::MAX, 1, 0));
-        assert_eq!(geometry.content, Rect::new(u16::MAX, u16::MAX, 1, 0));
-        assert!(
-            geometry
-                .panes
-                .values()
-                .all(|rect| rect.x == u16::MAX && rect.y == u16::MAX)
-        );
-    }
-
-    #[test]
-    fn snapshot_commit_repairs_removed_tab_and_pane_selection() {
-        let initial = layout(
-            vec![
-                Tab {
-                    tab_id: 1,
-                    root: Node::Leaf { pane_id: 1 },
-
-                    title: None,
-                },
-                Tab {
-                    tab_id: 2,
-                    root: Node::Leaf { pane_id: 2 },
-
-                    title: None,
-                },
-            ],
-            &[(1, 2, 2), (2, 2, 2)],
-        );
-        let mut tui = MultiPaneTui::new(initial).expect("valid layout");
-        tui.select_tab(2).expect("select second tab");
-
-        let mut committed = layout(
-            vec![Tab {
-                tab_id: 1,
-                root: Node::Leaf { pane_id: 1 },
-
-                title: None,
-            }],
-            &[(1, 2, 2)],
-        );
-        committed.revision = 2;
-        tui.apply_snapshot(committed).expect("valid commit");
-
-        assert_eq!(tui.current_tab(), 1);
-        assert_eq!(tui.focused_pane(), 1);
-    }
-
-    #[test]
-    fn pane_chord_consumes_commands_and_uses_focused_rect_aspect() {
-        let mut tui = MultiPaneTui::new(split_layout()).expect("valid layout");
-        let area = Rect::new(0, 0, 80, 24);
-
-        assert_eq!(
-            tui.handle_key(
-                KeyEvent::new(KeyCode::Char('p'), KeyModifiers::CONTROL),
-                area,
-            ),
-            KeyHandling::Consumed(vec![])
-        );
-        assert_eq!(tui.chord_mode(), ChordMode::Pane);
-        assert_eq!(
-            tui.handle_key(KeyEvent::new(KeyCode::Char('n'), KeyModifiers::NONE), area),
-            KeyHandling::Consumed(vec![UiIntent::CreatePane {
-                target_pane_id: 1,
-                axis: Axis::LeftRight,
-                position: NewPanePosition::Second,
-                grid_rows: 20,
-                grid_cols: 38,
-            }])
-        );
-        assert_eq!(tui.chord_mode(), ChordMode::None);
-
-        let _ = tui.handle_key(
-            KeyEvent::new(KeyCode::Char('p'), KeyModifiers::CONTROL),
-            area,
-        );
-        assert_eq!(
-            tui.handle_key(KeyEvent::new(KeyCode::Right, KeyModifiers::NONE), area),
-            KeyHandling::Consumed(vec![UiIntent::FocusPane { pane_id: 2 }])
-        );
-        assert_eq!(tui.focused_pane(), 2);
-    }
-
-    #[test]
-    fn option_arrows_focus_in_normal_and_pane_modes_without_forwarding_at_edges() {
-        let mut tui = MultiPaneTui::new(split_layout()).expect("valid layout");
-        let area = Rect::new(0, 0, 80, 24);
-
-        assert_eq!(
-            tui.handle_key(KeyEvent::new(KeyCode::Right, KeyModifiers::ALT), area),
-            KeyHandling::Consumed(vec![UiIntent::FocusPane { pane_id: 2 }])
-        );
-        assert_eq!(tui.chord_mode(), ChordMode::None);
-        assert_eq!(
-            tui.handle_key(KeyEvent::new(KeyCode::Right, KeyModifiers::ALT), area),
-            KeyHandling::Consumed(vec![]),
-            "an edge Option-arrow is still consumed"
-        );
-
-        let _ = tui.handle_key(
-            KeyEvent::new(KeyCode::Char('p'), KeyModifiers::CONTROL),
-            area,
-        );
-        assert_eq!(
-            tui.handle_key(KeyEvent::new(KeyCode::Down, KeyModifiers::ALT), area),
-            KeyHandling::Consumed(vec![UiIntent::FocusPane { pane_id: 3 }])
-        );
-        assert_eq!(tui.chord_mode(), ChordMode::Pane);
-    }
-
-    #[test]
-    fn ctrl_s_toggles_the_share_modal_and_leaves_any_chord_mode() {
-        let mut tui = MultiPaneTui::new(split_layout()).expect("valid layout");
-        let area = Rect::new(0, 0, 80, 24);
-
-        let _ = tui.handle_key(
-            KeyEvent::new(KeyCode::Char('p'), KeyModifiers::CONTROL),
-            area,
-        );
-        assert_eq!(
-            tui.handle_key(
-                KeyEvent::new(KeyCode::Char('s'), KeyModifiers::CONTROL),
-                area
-            ),
-            KeyHandling::Consumed(vec![]),
-            "share is a mux command and never reaches the PTY"
-        );
-        assert!(tui.share_open());
-        assert_eq!(tui.chord_mode(), ChordMode::None);
-
-        let _ = tui.handle_key(
-            KeyEvent::new(KeyCode::Char('s'), KeyModifiers::CONTROL),
-            area,
-        );
-        assert!(!tui.share_open(), "Ctrl+S closes what it opened");
-    }
-
-    #[test]
-    fn share_modal_claims_each_copy_once_and_closes_on_escape() {
-        let mut tui = MultiPaneTui::new(split_layout()).expect("valid layout");
-        let area = Rect::new(0, 0, 80, 24);
-        let _ = tui.handle_key(
-            KeyEvent::new(KeyCode::Char('s'), KeyModifiers::CONTROL),
-            area,
-        );
-
-        let _ = tui.handle_key(KeyEvent::new(KeyCode::Enter, KeyModifiers::NONE), area);
-        assert_eq!(tui.take_share_copy_request(), Some(ShareCopy::Ticket));
-        assert_eq!(
-            tui.take_share_copy_request(),
-            None,
-            "a claimed request must not copy again on the next key"
-        );
-
-        let _ = tui.handle_key(KeyEvent::new(KeyCode::Char('c'), KeyModifiers::NONE), area);
-        assert_eq!(tui.take_share_copy_request(), Some(ShareCopy::Code));
-
-        let _ = tui.handle_key(KeyEvent::new(KeyCode::Esc, KeyModifiers::NONE), area);
-        assert!(!tui.share_open());
-    }
-
-    #[test]
-    fn plain_share_key_reaches_the_pty_without_control() {
-        let mut tui = MultiPaneTui::new(split_layout()).expect("valid layout");
-        let area = Rect::new(0, 0, 80, 24);
-
-        assert_eq!(
-            tui.handle_key(KeyEvent::new(KeyCode::Char('s'), KeyModifiers::NONE), area),
-            KeyHandling::Forward
-        );
-        assert!(!tui.share_open());
-    }
-
-    #[test]
-    fn option_arrows_accept_extra_modifiers_but_not_control() {
-        let mut tui = MultiPaneTui::new(split_layout()).expect("valid layout");
-        let area = Rect::new(0, 0, 80, 24);
-
-        assert_eq!(
-            tui.handle_key(
-                KeyEvent::new(KeyCode::Right, KeyModifiers::ALT | KeyModifiers::SHIFT),
-                area,
-            ),
-            KeyHandling::Consumed(vec![UiIntent::FocusPane { pane_id: 2 }])
-        );
-        assert_eq!(
-            tui.handle_key(
-                KeyEvent::new(KeyCode::Left, KeyModifiers::ALT | KeyModifiers::CONTROL),
-                area,
-            ),
-            KeyHandling::Forward
-        );
-        assert_eq!(tui.focused_pane(), 2);
-    }
-
-    #[test]
-    fn esc_then_arrow_uses_option_focus_and_expired_esc_keeps_its_prior_behavior() {
-        let mut tui = MultiPaneTui::new(split_layout()).expect("valid layout");
-        let area = Rect::new(0, 0, 80, 24);
-        let now = Instant::now();
-        let mut pending_escape = PendingEscape::default();
-
-        pending_escape.start(now);
-        let option_arrow = pending_escape
-            .take_option_arrow(KeyEvent::new(KeyCode::Right, KeyModifiers::NONE))
-            .expect("Esc followed by an arrow becomes Option-arrow focus");
-        assert_eq!(
-            tui.handle_key(option_arrow, area),
-            KeyHandling::Consumed(vec![UiIntent::FocusPane { pane_id: 2 }])
-        );
-
-        let _ = tui.handle_key(
-            KeyEvent::new(KeyCode::Char('p'), KeyModifiers::CONTROL),
-            area,
-        );
-        pending_escape.start(now);
-        assert!(pending_escape.take_if_expired(now + ESC_PREFIX_WINDOW));
-        assert_eq!(
-            tui.handle_key(KeyEvent::new(KeyCode::Esc, KeyModifiers::NONE), area),
-            KeyHandling::Consumed(vec![])
-        );
-        assert_eq!(tui.chord_mode(), ChordMode::None);
-    }
-
-    #[test]
-    fn pane_chord_directional_splits_select_axis_and_child_position() {
-        let area = Rect::new(0, 0, 80, 24);
-        for (key, axis, position) in [
-            ('r', Axis::LeftRight, NewPanePosition::Second),
-            ('l', Axis::LeftRight, NewPanePosition::First),
-            ('d', Axis::TopBottom, NewPanePosition::Second),
-            ('u', Axis::TopBottom, NewPanePosition::First),
-        ] {
-            let mut tui = MultiPaneTui::new(split_layout()).expect("valid layout");
-            let _ = tui.handle_key(
-                KeyEvent::new(KeyCode::Char('p'), KeyModifiers::CONTROL),
-                area,
-            );
-            assert_eq!(
-                tui.handle_key(KeyEvent::new(KeyCode::Char(key), KeyModifiers::NONE), area),
-                KeyHandling::Consumed(vec![UiIntent::CreatePane {
-                    target_pane_id: 1,
-                    axis,
-                    position,
-                    grid_rows: 20,
-                    grid_cols: 38,
-                }]),
-                "key: {key}"
-            );
-            assert_eq!(tui.focused_pane(), 1, "key: {key}");
-            assert_eq!(tui.chord_mode(), ChordMode::None, "key: {key}");
-        }
-    }
-
-    #[test]
-    fn pane_lock_chord_toggles_lock_and_exits_mode() {
-        let area = Rect::new(0, 0, 80, 24);
-        let mut tui = MultiPaneTui::new(split_layout()).expect("valid layout");
-
-        let _ = tui.handle_key(
-            KeyEvent::new(KeyCode::Char('p'), KeyModifiers::CONTROL),
-            area,
-        );
-        assert_eq!(
-            tui.handle_key(KeyEvent::new(KeyCode::Char('k'), KeyModifiers::NONE), area),
-            KeyHandling::Consumed(vec![UiIntent::SetPaneLock {
-                pane_id: 1,
-                locked: true,
-            }])
-        );
-        assert_eq!(tui.chord_mode(), ChordMode::None);
-        assert!(is_chord_command(
-            ChordMode::Pane,
-            KeyEvent::new(KeyCode::Char('k'), KeyModifiers::NONE)
-        ));
-        assert!(!is_chord_navigation(
-            ChordMode::Pane,
-            KeyEvent::new(KeyCode::Char('k'), KeyModifiers::NONE)
-        ));
-    }
-
-    #[test]
-    fn pane_commands_exit_mode_even_when_no_intent_is_available() {
-        let area = Rect::new(0, 0, 80, 24);
-        let mut tui = MultiPaneTui::new(split_layout()).expect("valid layout");
-        tui.snapshot.panes.clear();
-
-        let _ = tui.handle_key(
-            KeyEvent::new(KeyCode::Char('p'), KeyModifiers::CONTROL),
-            area,
-        );
-        assert_eq!(
-            tui.handle_key(KeyEvent::new(KeyCode::Char('k'), KeyModifiers::NONE), area),
-            KeyHandling::Consumed(vec![])
-        );
-        assert_eq!(tui.chord_mode(), ChordMode::None);
-    }
-
-    #[test]
-    fn sticky_pane_mode_moves_focus_across_multiple_arrows() {
-        let mut tui = MultiPaneTui::new(split_layout()).expect("valid layout");
-        let area = Rect::new(0, 0, 80, 24);
-
-        let _ = tui.handle_key(
-            KeyEvent::new(KeyCode::Char('p'), KeyModifiers::CONTROL),
-            area,
-        );
-        assert_eq!(
-            tui.handle_key(KeyEvent::new(KeyCode::Right, KeyModifiers::NONE), area),
-            KeyHandling::Consumed(vec![UiIntent::FocusPane { pane_id: 2 }])
-        );
-        assert_eq!(tui.chord_mode(), ChordMode::Pane);
-        assert_eq!(
-            tui.handle_key(KeyEvent::new(KeyCode::Down, KeyModifiers::NONE), area),
-            KeyHandling::Consumed(vec![UiIntent::FocusPane { pane_id: 3 }])
-        );
-        assert_eq!(tui.chord_mode(), ChordMode::Pane);
-    }
-
-    #[test]
-    fn escape_clears_sticky_chord_mode_without_forwarding() {
-        let mut tui = MultiPaneTui::new(split_layout()).expect("valid layout");
-        let area = Rect::new(0, 0, 80, 24);
-
-        let _ = tui.handle_key(
-            KeyEvent::new(KeyCode::Char('p'), KeyModifiers::CONTROL),
-            area,
-        );
-        assert_eq!(
-            tui.handle_key(KeyEvent::new(KeyCode::Esc, KeyModifiers::NONE), area),
-            KeyHandling::Consumed(vec![])
-        );
-        assert_eq!(tui.chord_mode(), ChordMode::None);
-    }
-
-    #[test]
-    fn sticky_chord_mode_expires_two_seconds_after_last_key() {
-        let mut tui = MultiPaneTui::new(split_layout()).expect("valid layout");
-        let area = Rect::new(0, 0, 80, 24);
-        let now = Instant::now();
-        let _ = tui.handle_key(
-            KeyEvent::new(KeyCode::Char('p'), KeyModifiers::CONTROL),
-            area,
-        );
-        assert_eq!(tui.chord_mode(), ChordMode::Pane);
-        tui.chord_last_activity = now.checked_sub(CHORD_IDLE_TIMEOUT);
-        assert!(tui.expire_chord_mode(now));
-        assert_eq!(tui.chord_mode(), ChordMode::None);
-        assert!(tui.chord_last_activity.is_none());
-
-        let _ = tui.handle_key(
-            KeyEvent::new(KeyCode::Char('t'), KeyModifiers::CONTROL),
-            area,
-        );
-        assert_eq!(tui.chord_mode(), ChordMode::Tab);
-        tui.chord_last_activity = now.checked_sub(Duration::from_millis(1_999));
-        assert!(!tui.expire_chord_mode(now));
-        assert_eq!(tui.chord_mode(), ChordMode::Tab);
-        let _ = tui.handle_key(KeyEvent::new(KeyCode::Right, KeyModifiers::NONE), area);
-        tui.chord_last_activity = now.checked_sub(CHORD_IDLE_TIMEOUT);
-        assert!(tui.expire_chord_mode(now));
-        assert_eq!(tui.chord_mode(), ChordMode::None);
-    }
-
-    #[test]
-    fn forwarding_key_exits_sticky_pane_mode_once() {
-        let mut tui = MultiPaneTui::new(split_layout()).expect("valid layout");
-        let area = Rect::new(0, 0, 80, 24);
-
-        let _ = tui.handle_key(
-            KeyEvent::new(KeyCode::Char('p'), KeyModifiers::CONTROL),
-            area,
-        );
-        assert_eq!(
-            tui.handle_key(KeyEvent::new(KeyCode::Char('z'), KeyModifiers::NONE), area),
-            KeyHandling::Forward
-        );
-        assert_eq!(tui.chord_mode(), ChordMode::None);
-    }
-
-    #[test]
-    fn paste_exits_sticky_chord_mode_before_forwarding() {
-        let mut tui = MultiPaneTui::new(split_layout()).expect("valid layout");
-        let area = Rect::new(0, 0, 80, 24);
-        let _ = tui.handle_key(
-            KeyEvent::new(KeyCode::Char('p'), KeyModifiers::CONTROL),
-            area,
-        );
-
-        tui.exit_chord_mode();
-
-        assert_eq!(tui.chord_mode(), ChordMode::None);
-    }
-
-    #[test]
     fn member_labels_disambiguate_duplicate_display_names() {
         let members = vec![
             crate::layout::Member {
@@ -5533,436 +3916,5 @@ mod tests {
             "sam · aabbccdd"
         );
         assert_eq!(member_label(&members[2].peer_id, &members), "pat");
-    }
-
-    #[test]
-    fn initial_root_pane_grid_matches_its_bordered_viewport() {
-        let tui = MultiPaneTui::new(layout(
-            vec![Tab {
-                tab_id: 1,
-                root: Node::Leaf { pane_id: 1 },
-
-                title: None,
-            }],
-            &[(1, 1, 1)],
-        ))
-        .expect("valid layout");
-
-        let pane = tui
-            .geometry(Rect::new(0, 0, 80, 24))
-            .panes
-            .get(&1)
-            .copied()
-            .expect("root pane");
-        assert_eq!(grid_for_pane(pane), (20, 78));
-        assert_eq!(initial_root_pane_grid(80, 24), (20, 78));
-    }
-
-    #[test]
-    fn square_pane_create_chord_splits_top_to_bottom_without_a_tab_bar_gap() {
-        let mut tui = MultiPaneTui::new(layout(
-            vec![Tab {
-                tab_id: 1,
-                root: Node::Leaf { pane_id: 1 },
-
-                title: None,
-            }],
-            &[(1, 1, 1)],
-        ))
-        .expect("valid layout");
-        let area = Rect::new(0, 0, 12, 14);
-
-        let _ = tui.handle_key(
-            KeyEvent::new(KeyCode::Char('p'), KeyModifiers::CONTROL),
-            area,
-        );
-        assert_eq!(
-            tui.handle_key(KeyEvent::new(KeyCode::Char('n'), KeyModifiers::NONE), area),
-            KeyHandling::Consumed(vec![UiIntent::CreatePane {
-                target_pane_id: 1,
-                axis: Axis::TopBottom,
-                position: NewPanePosition::Second,
-                grid_rows: 10,
-                grid_cols: 10,
-            }])
-        );
-    }
-
-    #[test]
-    fn pane_delete_chord_targets_the_focused_pane() {
-        let mut tui = MultiPaneTui::new(split_layout()).expect("valid layout");
-        let area = Rect::new(0, 0, 80, 24);
-        let _ = tui.handle_key(
-            KeyEvent::new(KeyCode::Char('p'), KeyModifiers::CONTROL),
-            area,
-        );
-        let _ = tui.handle_key(KeyEvent::new(KeyCode::Right, KeyModifiers::NONE), area);
-        assert_eq!(tui.focused_pane(), 2);
-
-        let _ = tui.handle_key(
-            KeyEvent::new(KeyCode::Char('p'), KeyModifiers::CONTROL),
-            area,
-        );
-        assert_eq!(
-            tui.handle_key(KeyEvent::new(KeyCode::Char('x'), KeyModifiers::NONE), area),
-            KeyHandling::Consumed(vec![UiIntent::DeletePane { pane_id: 2 }])
-        );
-    }
-
-    #[test]
-    fn tall_pane_create_chord_uses_a_top_bottom_split_and_usable_grid() {
-        let mut tui = MultiPaneTui::new(layout(
-            vec![Tab {
-                tab_id: 1,
-                root: Node::Leaf { pane_id: 1 },
-
-                title: None,
-            }],
-            &[(1, 2, 2)],
-        ))
-        .expect("valid layout");
-        let area = Rect::new(0, 0, 20, 40);
-
-        let _ = tui.handle_key(
-            KeyEvent::new(KeyCode::Char('p'), KeyModifiers::CONTROL),
-            area,
-        );
-        assert_eq!(
-            tui.handle_key(KeyEvent::new(KeyCode::Char('n'), KeyModifiers::NONE), area),
-            KeyHandling::Consumed(vec![UiIntent::CreatePane {
-                target_pane_id: 1,
-                axis: Axis::TopBottom,
-                position: NewPanePosition::Second,
-                grid_rows: 36,
-                grid_cols: 18,
-            }])
-        );
-    }
-
-    #[test]
-    fn forwarding_keys_exit_a_chord_mode_and_are_forwarded() {
-        let mut tui = MultiPaneTui::new(split_layout()).expect("valid layout");
-        let area = Rect::new(0, 0, 80, 24);
-
-        for prefix in ['p', 't'] {
-            assert_eq!(
-                tui.handle_key(
-                    KeyEvent::new(KeyCode::Char(prefix), KeyModifiers::CONTROL),
-                    area
-                ),
-                KeyHandling::Consumed(vec![])
-            );
-            assert_eq!(
-                tui.handle_key(KeyEvent::new(KeyCode::Char('z'), KeyModifiers::NONE), area),
-                KeyHandling::Forward
-            );
-            assert_eq!(tui.chord_mode(), ChordMode::None);
-        }
-    }
-
-    #[test]
-    fn pane_focus_uses_the_nearest_directional_leaf_and_stops_at_edges() {
-        let mut tui = MultiPaneTui::new(split_layout()).expect("valid layout");
-        let area = Rect::new(0, 0, 80, 24);
-        let _ = tui.handle_key(
-            KeyEvent::new(KeyCode::Char('p'), KeyModifiers::CONTROL),
-            area,
-        );
-        assert_eq!(
-            tui.handle_key(KeyEvent::new(KeyCode::Right, KeyModifiers::NONE), area),
-            KeyHandling::Consumed(vec![UiIntent::FocusPane { pane_id: 2 }])
-        );
-        let _ = tui.handle_key(
-            KeyEvent::new(KeyCode::Char('p'), KeyModifiers::CONTROL),
-            area,
-        );
-        assert_eq!(
-            tui.handle_key(KeyEvent::new(KeyCode::Down, KeyModifiers::NONE), area),
-            KeyHandling::Consumed(vec![UiIntent::FocusPane { pane_id: 3 }])
-        );
-        let _ = tui.handle_key(
-            KeyEvent::new(KeyCode::Char('p'), KeyModifiers::CONTROL),
-            area,
-        );
-        assert_eq!(
-            tui.handle_key(KeyEvent::new(KeyCode::Left, KeyModifiers::NONE), area),
-            KeyHandling::Consumed(vec![UiIntent::FocusPane { pane_id: 1 }])
-        );
-        assert_eq!(tui.focused_pane(), 1);
-        let _ = tui.handle_key(
-            KeyEvent::new(KeyCode::Char('p'), KeyModifiers::CONTROL),
-            area,
-        );
-        assert_eq!(
-            tui.handle_key(KeyEvent::new(KeyCode::Up, KeyModifiers::NONE), area),
-            KeyHandling::Consumed(vec![UiIntent::FocusPane { pane_id: 2 }])
-        );
-        assert_eq!(tui.focused_pane(), 2);
-
-        let mut edge_tui = MultiPaneTui::new(split_layout()).expect("valid layout");
-        let _ = edge_tui.handle_key(
-            KeyEvent::new(KeyCode::Char('p'), KeyModifiers::CONTROL),
-            area,
-        );
-        assert_eq!(
-            edge_tui.handle_key(KeyEvent::new(KeyCode::Left, KeyModifiers::NONE), area),
-            KeyHandling::Consumed(vec![])
-        );
-        assert_eq!(edge_tui.focused_pane(), 1);
-
-        let _ = edge_tui.handle_key(
-            KeyEvent::new(KeyCode::Char('p'), KeyModifiers::CONTROL),
-            area,
-        );
-        let _ = edge_tui.handle_key(KeyEvent::new(KeyCode::Right, KeyModifiers::NONE), area);
-        assert_eq!(edge_tui.focused_pane(), 2);
-        let _ = edge_tui.handle_key(
-            KeyEvent::new(KeyCode::Char('p'), KeyModifiers::CONTROL),
-            area,
-        );
-        let _ = edge_tui.handle_key(KeyEvent::new(KeyCode::Down, KeyModifiers::NONE), area);
-        assert_eq!(edge_tui.focused_pane(), 3);
-        for key in [KeyCode::Down, KeyCode::Right] {
-            let _ = edge_tui.handle_key(
-                KeyEvent::new(KeyCode::Char('p'), KeyModifiers::CONTROL),
-                area,
-            );
-            assert_eq!(
-                edge_tui.handle_key(KeyEvent::new(key, KeyModifiers::NONE), area),
-                KeyHandling::Consumed(vec![])
-            );
-            assert_eq!(edge_tui.focused_pane(), 3);
-        }
-    }
-
-    #[test]
-    fn creator_selects_only_its_explicitly_reserved_tab_after_that_tab_commits() {
-        let mut tui = MultiPaneTui::new(layout(
-            vec![Tab {
-                tab_id: 1,
-                root: Node::Leaf { pane_id: 1 },
-
-                title: None,
-            }],
-            &[(1, 2, 2)],
-        ))
-        .expect("valid layout");
-
-        tui.select_created_tab(2);
-        let mut unrelated = layout(
-            vec![
-                Tab {
-                    tab_id: 1,
-                    root: Node::Leaf { pane_id: 1 },
-
-                    title: None,
-                },
-                Tab {
-                    tab_id: 3,
-                    root: Node::Leaf { pane_id: 3 },
-
-                    title: None,
-                },
-            ],
-            &[(1, 2, 2), (3, 2, 2)],
-        );
-        unrelated.revision = 2;
-        tui.apply_snapshot(unrelated).expect("unrelated commit");
-        assert_eq!(tui.current_tab(), 1);
-
-        let mut targeted = layout(
-            vec![
-                Tab {
-                    tab_id: 1,
-                    root: Node::Leaf { pane_id: 1 },
-
-                    title: None,
-                },
-                Tab {
-                    tab_id: 3,
-                    root: Node::Leaf { pane_id: 3 },
-
-                    title: None,
-                },
-                Tab {
-                    tab_id: 2,
-                    root: Node::Leaf { pane_id: 2 },
-
-                    title: None,
-                },
-            ],
-            &[(1, 2, 2), (2, 2, 2), (3, 2, 2)],
-        );
-        targeted.revision = 3;
-        tui.apply_snapshot(targeted).expect("targeted tab commit");
-
-        assert_eq!(tui.current_tab(), 2);
-        assert_eq!(tui.focused_pane(), 2);
-    }
-
-    #[test]
-    fn tab_chord_switches_and_creates_or_deletes_tabs_without_forwarding_keys() {
-        let mut tui = MultiPaneTui::new(layout(
-            vec![
-                Tab {
-                    tab_id: 1,
-                    root: Node::Leaf { pane_id: 1 },
-
-                    title: None,
-                },
-                Tab {
-                    tab_id: 2,
-                    root: Node::Leaf { pane_id: 2 },
-
-                    title: None,
-                },
-            ],
-            &[(1, 2, 2), (2, 2, 2)],
-        ))
-        .expect("valid layout");
-        let area = Rect::new(0, 0, 12, 8);
-
-        let _ = tui.handle_key(
-            KeyEvent::new(KeyCode::Char('t'), KeyModifiers::CONTROL),
-            area,
-        );
-        assert_eq!(
-            tui.handle_key(KeyEvent::new(KeyCode::Right, KeyModifiers::NONE), area),
-            KeyHandling::Consumed(vec![UiIntent::SwitchTab { tab_id: 2 }])
-        );
-        assert_eq!(
-            tui.handle_key(
-                KeyEvent::new(KeyCode::Char('t'), KeyModifiers::CONTROL),
-                area,
-            ),
-            KeyHandling::Consumed(vec![])
-        );
-        assert_eq!(
-            tui.handle_key(KeyEvent::new(KeyCode::Char('n'), KeyModifiers::NONE), area),
-            KeyHandling::Consumed(vec![UiIntent::CreateTab {
-                grid_rows: 4,
-                grid_cols: 10,
-            }])
-        );
-        assert_eq!(tui.chord_mode(), ChordMode::None);
-
-        let _ = tui.handle_key(
-            KeyEvent::new(KeyCode::Char('t'), KeyModifiers::CONTROL),
-            area,
-        );
-        assert_eq!(
-            tui.handle_key(KeyEvent::new(KeyCode::Char('x'), KeyModifiers::NONE), area),
-            KeyHandling::Consumed(vec![UiIntent::DeleteTab { tab_id: 2 }])
-        );
-        assert_eq!(tui.chord_mode(), ChordMode::None);
-    }
-
-    #[test]
-    fn sticky_tab_mode_switches_tabs_repeatedly() {
-        let mut tui = MultiPaneTui::new(layout(
-            vec![
-                Tab {
-                    tab_id: 1,
-                    root: Node::Leaf { pane_id: 1 },
-
-                    title: None,
-                },
-                Tab {
-                    tab_id: 2,
-                    root: Node::Leaf { pane_id: 2 },
-
-                    title: None,
-                },
-                Tab {
-                    tab_id: 3,
-                    root: Node::Leaf { pane_id: 3 },
-
-                    title: None,
-                },
-            ],
-            &[(1, 2, 2), (2, 2, 2), (3, 2, 2)],
-        ))
-        .expect("valid layout");
-        let area = Rect::new(0, 0, 12, 8);
-
-        let _ = tui.handle_key(
-            KeyEvent::new(KeyCode::Char('t'), KeyModifiers::CONTROL),
-            area,
-        );
-        assert_eq!(
-            tui.handle_key(KeyEvent::new(KeyCode::Right, KeyModifiers::NONE), area),
-            KeyHandling::Consumed(vec![UiIntent::SwitchTab { tab_id: 2 }])
-        );
-        assert_eq!(tui.chord_mode(), ChordMode::Tab);
-        assert_eq!(
-            tui.handle_key(KeyEvent::new(KeyCode::Right, KeyModifiers::NONE), area),
-            KeyHandling::Consumed(vec![UiIntent::SwitchTab { tab_id: 3 }])
-        );
-        assert_eq!(tui.chord_mode(), ChordMode::Tab);
-    }
-
-    #[test]
-    fn normal_keys_escape_and_function_keys_leave_f9_and_f10_for_the_pty() {
-        let mut tui = MultiPaneTui::new(split_layout()).expect("valid layout");
-        let area = Rect::new(0, 0, 80, 24);
-        assert_eq!(
-            tui.handle_key(KeyEvent::new(KeyCode::Char('z'), KeyModifiers::NONE), area),
-            KeyHandling::Forward
-        );
-        assert_eq!(
-            tui.handle_key(KeyEvent::new(KeyCode::Esc, KeyModifiers::NONE), area),
-            KeyHandling::Forward
-        );
-        let _ = tui.handle_key(
-            KeyEvent::new(KeyCode::Char('p'), KeyModifiers::CONTROL),
-            area,
-        );
-        assert_eq!(
-            tui.handle_key(KeyEvent::new(KeyCode::Esc, KeyModifiers::NONE), area),
-            KeyHandling::Consumed(vec![])
-        );
-        assert_eq!(
-            tui.handle_key(KeyEvent::new(KeyCode::F(9), KeyModifiers::NONE), area),
-            KeyHandling::Forward
-        );
-        assert_eq!(
-            tui.handle_key(KeyEvent::new(KeyCode::F(10), KeyModifiers::NONE), area),
-            KeyHandling::Forward
-        );
-        assert_eq!(
-            tui.handle_key(
-                KeyEvent::new(KeyCode::Char('q'), KeyModifiers::CONTROL),
-                area,
-            ),
-            KeyHandling::Quit
-        );
-    }
-
-    #[test]
-    fn shift_l_toggles_the_session_lock_from_whatever_state_it_is_in() {
-        let area = Rect::new(0, 0, 80, 24);
-        let mut tui = MultiPaneTui::new(split_layout()).expect("valid layout");
-
-        let _ = tui.handle_key(
-            KeyEvent::new(KeyCode::Char('p'), KeyModifiers::CONTROL),
-            area,
-        );
-        assert_eq!(
-            tui.handle_key(KeyEvent::new(KeyCode::Char('L'), KeyModifiers::SHIFT), area),
-            KeyHandling::Consumed(vec![UiIntent::SetSessionLock { locked: true }]),
-            "Shift+L should offer to lock a session that is currently open"
-        );
-        assert_eq!(tui.chord_mode(), ChordMode::None);
-
-        tui.set_session_locked(true);
-        let _ = tui.handle_key(
-            KeyEvent::new(KeyCode::Char('p'), KeyModifiers::CONTROL),
-            area,
-        );
-        assert_eq!(
-            tui.handle_key(KeyEvent::new(KeyCode::Char('L'), KeyModifiers::SHIFT), area),
-            KeyHandling::Consumed(vec![UiIntent::SetSessionLock { locked: false }]),
-            "and to unlock one that is locked"
-        );
     }
 }
